+++ doc/changes.rst
@@ -5,14 +5,18 @@ Change History
 3.6.11-r1
 =========
 
+You can now use the hot backup functionality introduced in SQLite
+3.6.11.
+
 Updated a VFS test to reflect changes in SQLite underlying error
 handling.  (Previously SQLite almost always returned :exc:`FullError`
 on any write that had an error but now returns :exc:`SQLiteError`.)
 
-You can now use the hot backup functionality introduced in SQLite
-3.6.11.
-
 Changed close methods so that Connections can be released earlier.
+
+In prior releases a :meth:`closed <Cursor.close>` could still be used
+(reincarnated).  That is no longer the case and you will get
+:exc:`CursorClosedError`.
 
 3.6.10-r1
 =========